requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai>=1.0.0
